Extract a bunch of query objects for progress reports
