+++ app/controllers/teachers/progress_reports/activity_sessions_controller.rb
@@ -3,15 +3,15 @@ class Teachers::ProgressReports::ActivitySessionsController < Teachers::Progress
     respond_to do |format|
       format.html
       format.json do
-        query = ActivitySession.for_standalone_progress_report(current_user, params)
+        query = ProgressReports::ActivitySession.new(current_user).results(params)
         page_count = (query.count / ActivitySession::RESULTS_PER_PAGE.to_f).ceil
         activity_sessions = query.paginate(params[:page], ActivitySession::RESULTS_PER_PAGE)
         activity_session_json = activity_sessions.map do |activity_session|
           ::ProgressReports::ActivitySessionSerializer.new(activity_session).as_json(root: false)
         end
-        classrooms = Classroom.for_standards_report(current_user, {})
-        students = User.for_standards_report(current_user, {})
-        units = Unit.for_standards_progress_report(current_user, {})
+        classrooms = ProgressReports::Standards::Classroom.new(current_user).results({})
+        students = ProgressReports::Standards::Student.new(current_user).results({})
+        units = ProgressReports::Standards::Unit.new(current_user).results({})
         render json: {
           activity_sessions: activity_session_json,
           classrooms: classrooms,
