+++ app/controllers/teachers/progress_reports/standards/classroom_students_controller.rb
@@ -3,7 +3,7 @@ class Teachers::ProgressReports::Standards::ClassroomStudentsController < Teache
     respond_to do |format|
       format.html
       format.json do
-        students = User.for_standards_report(current_user, params)
+        students = ::ProgressReports::Standards::Student.new(current_user).results(params)
         students_json = students.map do |student|
           serializer = ::ProgressReports::Standards::StudentSerializer.new(student)
           # Doing this because can't figure out how to get custom params into serializers
@@ -13,7 +13,7 @@ class Teachers::ProgressReports::Standards::ClassroomStudentsController < Teache
         render json: {
           students: students_json,
           classroom: current_user.classrooms.find(params[:classroom_id]),
-          units: Unit.for_standards_progress_report(current_user, {}),
+          units: ProgressReports::Standards::Unit.new(current_user).results({}),
           teacher: UserWithEmailSerializer.new(current_user).as_json(root: false)
         }
       end
