+++ app/controllers/teachers/progress_reports/standards/classroom_topics_controller.rb
@@ -3,7 +3,7 @@ class Teachers::ProgressReports::Standards::ClassroomTopicsController < Teachers
     respond_to do |format|
       format.html
       format.json do
-        topics = Topic.for_standards_report(current_user, params)
+        topics = ::ProgressReports::Standards::Topic.new(current_user).results(params)
         topics_json = topics.map do |topic|
           serializer = ::ProgressReports::Standards::TopicSerializer.new(topic)
           # Doing this because can't figure out how to get custom params into serializers
@@ -12,7 +12,7 @@ class Teachers::ProgressReports::Standards::ClassroomTopicsController < Teachers
         end
         render json: {
           topics: topics_json,
-          units: Unit.for_standards_progress_report(current_user, {}),
+          units: ProgressReports::Standards::Unit.new(current_user).results({}),
           classroom: current_user.classrooms.find(params[:classroom_id]),
           teacher: UserWithEmailSerializer.new(current_user).as_json(root: false)
         }
