+++ app/controllers/teachers/progress_reports/standards/classrooms_controller.rb
@@ -3,7 +3,7 @@ class Teachers::ProgressReports::Standards::ClassroomsController < Teachers::Pro
     respond_to do |format|
       format.html
       format.json do
-        classrooms = Classroom.for_standards_report(current_user, params)
+        classrooms = ProgressReports::Standards::Classroom.new(current_user).results(params)
         classroom_json = classrooms.map do |classroom|
           ::ProgressReports::Standards::ClassroomSerializer.new(classroom).as_json(root: false)
         end
