+++ app/models/activity_session.rb
@@ -37,19 +37,17 @@ class ActivitySession < ActiveRecord::Base
 
   RESULTS_PER_PAGE = 25
 
-  def self.for_standalone_progress_report(teacher, filters)
-    filters = (filters || {}).with_indifferent_access
-    query = includes(:user, :activity => [:topic, :classification], :classroom_activity => :classroom)
-      .references(:classification)
-      .completed
-      .by_teacher(teacher)
-      .order(search_sort_sql(filters[:sort]))
-    with_filters(query, filters)
-  end
-
   def self.paginate(current_page, per_page)
     offset = (current_page.to_i - 1) * per_page
     limit(per_page).offset(offset)
+  end
+
+  def self.with_best_scores
+    where(is_final_score: true)
+  end
+
+  def self.by_teacher(teacher)
+    self.joins(:user => :teacher).where(teachers_users: {id: teacher.id})
   end
 
   def self.with_filters(query, filters)
@@ -74,26 +72,6 @@ class ActivitySession < ActiveRecord::Base
     end
 
     query
-  end
-
-  def self.for_standards_report(teacher, filters)
-    query = select(<<-SELECT
-      activity_sessions.*,
-      activities.topic_id as topic_id
-    SELECT
-    ).completed
-      .with_best_scores
-      .by_teacher(teacher)
-    query = with_filters(query, filters)
-    query
-  end
-
-  def self.with_best_scores
-    where(is_final_score: true)
-  end
-
-  def self.by_teacher(teacher)
-    self.joins(:user => :teacher).where(teachers_users: {id: teacher.id})
   end
 
   def determine_if_final_score
