+++ app/models/classroom.rb
@@ -29,27 +29,10 @@ class Classroom < ActiveRecord::Base
     end
   end
 
-  def self.for_standards_report(teacher, filters)
-    with(user_info: User.for_standards_report(teacher, filters))
-      .select(<<-SQL
-        classrooms.name as name,
-        classrooms.id as id,
-        classrooms.teacher_id,
-        COUNT(DISTINCT(user_info.id)) as total_student_count,
-        SUM(CASE WHEN user_info.average_score > 0.75 THEN 1 ELSE 0 END) as proficient_student_count,
-        SUM(CASE WHEN user_info.average_score >= 0.50 AND user_info.average_score <= 0.75 THEN 1 ELSE 0 END) as near_proficient_student_count,
-        SUM(CASE WHEN user_info.average_score < 0.50 THEN 1 ELSE 0 END) as not_proficient_student_count
-      SQL
-      )
-      .joins(:students)
-      .joins('INNER JOIN user_info ON user_info.id = users.id')
-      .order("classrooms.name asc")
-      .group("classrooms.id")
-  end
-
   def unique_topic_count
     filters = {}
-    ActivitySession.from_cte('best_activity_sessions', ActivitySession.for_standards_report(teacher, filters))
+    best_activity_sessions = ProgressReports::Standards::ActivitySession.new(teacher).results(filters)
+    ActivitySession.from_cte('best_activity_sessions', best_activity_sessions)
       .select("COUNT(DISTINCT(activities.topic_id)) as topic_count")
       .joins('JOIN activities ON activities.id = best_activity_sessions.activity_id')
       .joins('JOIN classroom_activities ON classroom_activities.id = best_activity_sessions.classroom_activity_id')
