+++ app/models/csv_exporter/standards/topic_student.rb
@@ -22,9 +22,8 @@ module CsvExporter::Standards
     end
 
     def model_data(teacher, filters)
-      ::User.for_standards_report(
-        teacher,
-        HashWithIndifferentAccess.new(filters) || {})
+      ::ProgressReports::Standards::Student.new(teacher)
+        .results(HashWithIndifferentAccess.new(filters) || {})
     end
 
     private
