+++ spec/models/activity_session_spec.rb
@@ -79,82 +79,6 @@ describe ActivitySession, :type => :model do
 
   end
 
-  describe "#for_standards_report" do
-    include_context 'Topic Progress Report'
-    let(:filters) { {} }
-    subject { ActivitySession.for_standards_report(teacher, filters)}
-
-    it "must retrieve completed activity sessions representing the best scores for a teacher's students" do
-      expect(subject.size).to eq(best_activity_sessions.size)
-    end
-  end
-
-  describe "#for_standalone_progress_report" do
-    include_context 'Topic Progress Report'
-
-    subject { ActivitySession.for_standalone_progress_report(teacher, filters).to_a }
-
-    context 'sorting' do
-      before do
-        Timecop.freeze
-      end
-
-      after do
-        Timecop.return
-      end
-
-      context 'by default' do
-        let(:filters) { {} }
-        it 'sorts by completed_at descending' do
-          expect(subject.first.completed_at).to be_within(1.second).of fred_first_grade_topic_session.completed_at
-        end
-      end
-
-      context 'by activity classification' do
-        let(:filters) { {sort: {field: 'activity_classification_name', direction: 'asc'} } }
-        it 'retrieves results in the appropriate order' do
-          # Primary sort by classification name, secondary by student name
-          expect(subject.first.user.name).to eq(alice.name)
-        end
-      end
-
-      context 'by student name' do
-        let(:filters) { {sort: {field: 'student_name', direction: 'desc'} } }
-        it 'retrieves results in the appropriate order' do
-          expect(subject.first.user.name).to eq(zojirushi.name)
-        end
-      end
-
-      context 'by completion date' do
-        let(:filters) { {sort: {field: 'completed_at', direction: 'desc'} } }
-        it 'retrieves results in the appropriate order' do
-          expect(subject.first.completed_at).to be_within(1.second).of fred_first_grade_topic_session.completed_at
-        end
-      end
-
-      context 'by activity name' do
-        let(:filters) { {sort: {field: 'activity_name', direction: 'asc'} } }
-        it 'retrieves results in the appropriate order' do
-          expect(subject.first.activity.name).to eq(activity_for_first_grade_topic.name)
-        end
-      end
-
-      context 'by score' do
-        let(:filters) { {sort: {field: 'percentage', direction: 'desc'} } }
-        it 'retrieves results in the appropriate order' do
-          expect(subject.first.percentage).to eq(best_score_sessions.first.percentage)
-        end
-      end
-
-      context 'by standard' do
-        let(:filters) { {sort: {field: 'standard', direction: 'asc'} } }
-        it 'retrieves results in the appropriate order' do
-          expect(subject.first.activity.topic.name).to eq(first_grade_topic.name)
-        end
-      end
-    end
-  end
-
   describe "#by_teacher" do
     # This setup is very convoluted... the factories appear to be untrustworthy w/r/t generating extra records
     let!(:current_teacher) { FactoryGirl.create(:teacher) }
