+++ spec/serializers/progress_reports/standards/classroom_serializer_spec.rb
@@ -5,7 +5,7 @@ describe ProgressReports::Standards::ClassroomSerializer, type: :serializer do
   let!(:classroom) { FactoryGirl.create(:classroom_with_one_student, teacher: teacher) }
   let(:activity) { FactoryGirl.create(:activity) }
   let(:classroom_activity) { FactoryGirl.create(:classroom_activity, classroom: classroom, activity: activity) }
-  let(:classroom_for_report) { Classroom.for_standards_report(teacher, {}).first }
+  let(:classroom_for_report) { ProgressReports::Standards::Classroom.new(teacher).results({}).first }
   let(:serializer) { described_class.new(classroom_for_report) }
 
   before do
